+++ UvaSoftware.Scanii/ScaniiClients.cs
@@ -2,6 +2,7 @@ using System;
 using System.Net.Http;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
+using UvaSoftware.Scanii.Entities;
 using UvaSoftware.Scanii.Internal;
 
 namespace UvaSoftware.Scanii
@@ -11,20 +12,30 @@ namespace UvaSoftware.Scanii
     public static IScaniiClient CreateDefault(string key, string secret, ILogger logger = null,
       HttpClient client = null, ScaniiTarget target = null)
     {
+      if (secret == null) throw new ArgumentNullException(nameof(secret));
+
+      if (key == null) throw new ArgumentNullException(nameof(key));
+
+      if (key.Contains(":")) throw new ArgumentException("secret may not contains ':");
+
       logger ??= NullLogger.Instance;
       client ??= new HttpClient();
       target ??= ScaniiTarget.Auto;
-      ;
       return new DefaultScaniiClient(target, key, secret, logger, client);
     }
 
-    public static IScaniiClient CreateDefault(string authToken, ILogger logger = null,
+    public static IScaniiClient CreateDefault(ScaniiAuthToken authToken, ILogger logger = null,
       HttpClient client = null, ScaniiTarget target = null)
     {
+      if (authToken == null)
+      {
+        throw new ArgumentNullException(nameof(authToken));
+      }
+
       logger ??= NullLogger.Instance;
       client ??= new HttpClient();
       target ??= ScaniiTarget.Auto;
-      return new DefaultScaniiClient(target, authToken, "", logger, client);
+      return new DefaultScaniiClient(target, authToken.ResourceId, "", logger, client);
     }
   }
 }